% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{boundary}
\alias{boundary}
\title{Boundary of null hypothesis of forecast dominance}
\usage{
boundary(p, q, type)
}
\arguments{
\item{p}{forecast probabilities (numbers in [0,1]). The null hypothesis is
that \code{p} is a better forecast than \code{q}.}

\item{q}{forecast probabilities (numbers in [0,1]).}

\item{type}{scoring function. Currently available are \code{"brier"} for the
Brier score, \code{"log"} for the logarithmic score, \code{"spherical"}
for the spherical score.}
}
\value{
Vector of probabilities.
}
\description{
For two competing probability forecasts \code{p} and \code{q}, this function
computes the event probability under which \code{p} and \code{q} are
equally good with respect to a given scoring function.
}
\references{
Preprint on e-values for forecast evaluation:

A. Henzi and J. F Ziegel. Valid sequential inference on probability forecast
performance. arXiv preprint arXiv:2103.08402, 2021.

Proper scoring rules for probabilities:

Gneiting, Tilmann, and Adrian E. Raftery. "Strictly proper scoring rules,
prediction, and estimation." Journal of the American statistical Association
102.477 (2007): 359-378.
}
