% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_alpha}
\alias{check_alpha}
\title{Check signigicance level input}
\usage{
check_alpha(alpha)
}
\arguments{
\item{alpha}{significance level (should be single number in (0,1)).}
}
\description{
Check signigicance level input
}
\keyword{internal}
