% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_c}
\alias{check_c}
\title{Check condition input}
\usage{
check_c(c)
}
\arguments{
\item{c}{conditioning variable (should be vector of 0 and/or 1).}
}
\description{
Check condition input
}
\keyword{internal}
