% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_lengths}
\alias{check_lengths}
\title{Check equal vector lengths input}
\usage{
check_lengths(...)
}
\arguments{
\item{...}{vectors (should all have the same length).}
}
\description{
Check equal vector lengths input
}
\keyword{internal}
