% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_prob}
\alias{check_prob}
\title{Check probability input}
\usage{
check_prob(prob)
}
\arguments{
\item{prob}{probabilities (should be vector numbers in (0,1)).}
}
\description{
Check probability input
}
\keyword{internal}
