% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_stopped}
\alias{check_stopped}
\title{Check input for stopped argument}
\usage{
check_stopped(x)
}
\arguments{
\item{x}{whether e-value is stopped (should a single boolean).}
}
\description{
Check input for stopped argument
}
\keyword{internal}
