% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_y}
\alias{check_y}
\title{Check observation input}
\usage{
check_y(y)
}
\arguments{
\item{y}{observations (should be vector of 0 and/or 1).}
}
\description{
Check observation input
}
\keyword{internal}
