% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{etest}
\alias{etest}
\title{Hypothesis testing with e-values for sequential forecast dominance}
\usage{
etest(
  p,
  q,
  y,
  alt,
  c = rep(1, length(y)),
  type,
  h,
  stopped = FALSE,
  alpha = NULL
)
}
\arguments{
\item{p}{forecast probabilities (numbers in [0,1]). The null hypothesis is
that \code{p} is a better forecast than \code{q}.}

\item{q}{forecast probabilities (numbers in [0,1]).}

\item{y}{observations (0 or 1).}

\item{alt}{event probabilities under alternative hypothesis
(numbers in [0,1]).}

\item{c}{conditioning variable (0 or 1). E-values with \code{c = 0} will be
ignored (set to 1), and only e-values with \code{c = 1} are computed.}

\item{type}{scoring function. Currently available are \code{"brier"} for the
Brier score, \code{"log"} for the logarithmic score, \code{"spherical"}
for the spherical score, and \code{"dominance"} for forecast dominance
with respect to all scoring rules.}

\item{h}{forecast lag. For example, daily forecasts for the next day have lag
1, daily forecasts for an event two days ahead have lag 2.}

\item{stopped}{compute stopped e-values for the given level \code{alpha}?}

\item{alpha}{significance level (number in (0,1)). Has no effect if
\code{stopped} is \code{FALSE}.}
}
\value{
A list containing

\item{\code{evalue}}{e-value for the hypothesis test.}
\item{\code{pvalue}}{the corresponding p-value. If \code{stopped = FALSE},
  this is the anytime valid p-value, otherwise it is \code{1/evalue} up to a
  correction factor.}
\item{\code{alpha}}{significance level.}
\item{\code{h}}{forecast lag.}
\item{\code{stopped}}{whether the e-value is computed with optional stopping
 or not.}
\item{\code{stop_ind}}{the index at which the stopped e-value is reached.
  Equals the total sample size \code{n} if \code{stopped} is \code{FALSE}.}
\item{\code{n}}{total sample size (\code{length(y))}.}
\item{\code{alpha}}{significance level.}
}
\description{
For two competing probability forecasts \code{p} and \code{q} and alternative
probabilities \code{alt}, this function computes the e-value with the null
hypothesis that \code{p} is better than \code{q} and maximal power against
the alternative probability \code{alt}, under a given scoring function.
}
\details{
There is no default value for the alternative \code{alt}. The alternative
must be greater (less) than \code{\link{boundary}(p, q, type)} if
\code{q > p} (if \code{q < p}).

For testing forecast dominance (\code{type} equal to \code{"dominance"}),
rejecting the null hypothesis either means that \code{q} dominates \code{p}
with respect to all scoring rules, or that neither \code{p} dominates
\code{q} nor \code{q} dominates \code{p}.
}
\note{
Large e-values (greater than 1) give evidence against the null hypothesis and
indicate that \code{q} is a better forecast than \code{p}.

If the data (\code{p}, \code{q} or \code{y}) contains \code{NA}, these must
not be removed! It is assumed that the time lag is always \code{h} between
one entry of the data and the next, for example between \code{p[i]} and
\code{p[i + 1]}.
}
\references{
Preprint on e-values for forecast evaluation:

A. Henzi and J. F Ziegel. Valid sequential inference on probability forecast
performance. arXiv preprint arXiv:2103.08402, 2021.

Proper scoring rules for probabilities:

Gneiting, Tilmann, and Adrian E. Raftery. "Strictly proper scoring rules,
prediction, and estimation." Journal of the American statistical Association
102.477 (2007): 359-378.
}
