% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{evalue}
\alias{evalue}
\title{E-values for testing forecast dominance}
\usage{
evalue(y, p, q, alt, c = rep(1, length(y)), type)
}
\arguments{
\item{y}{observations (0 or 1).}

\item{p}{forecast probabilities (numbers in [0,1]). The null hypothesis is
that \code{p} is a better forecast than \code{q}.}

\item{q}{forecast probabilities (numbers in [0,1]).}

\item{alt}{event probabilities under alternative hypothesis
(numbers in [0,1]).}

\item{c}{conditioning variable (0 or 1). E-values with \code{c = 0} will be
ignored (set to 1), and only e-values with \code{c = 1} are computed.}

\item{type}{scoring function. Currently available are \code{"brier"} for the
Brier score, \code{"log"} for the logarithmic score, \code{"spherical"}
for the spherical score, and \code{"dominance"} for forecast dominance
with respect to all scoring rules.}
}
\value{
Vector of e-values. Large e-values (greater than 1) indicate evidence
against the null hypothesis that \code{p} is a better forecast than \code{q}.
}
\description{
For two competing probability forecasts \code{p} and \code{q} and an
alternative probability \code{alt}, this function computes for every time
point the e-value with the null hypothesis that \code{p} is better than
\code{q} and maximal power against the alternative probabilities \code{alt},
under a given scoring function.
}
\details{
For testing forecast dominance (\code{type} equal to \code{"dominance"}),
rejecting the null hypothesis either means that \code{q} dominates \code{p}
with respect to all scoring rules, or that neither \code{p} dominates
\code{q} nor \code{q} dominates \code{p}.
}
\references{
Preprint on e-values for forecast evaluation:

A. Henzi and J. F Ziegel. Valid sequential inference on probability forecast
performance. arXiv preprint arXiv:2103.08402, 2021.

Proper scoring rules for probabilities:

Gneiting, Tilmann, and Adrian E. Raftery. "Strictly proper scoring rules,
prediction, and estimation." Journal of the American statistical Association
102.477 (2007): 359-378.
}
