% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{score}
\alias{score}
\title{Consistent scoring functions for probabilities}
\usage{
score(y, p, type)
}
\arguments{
\item{y}{observations (0 or 1).}

\item{p}{forecast probabilities (numbers in [0,1]).}

\item{type}{scoring function to be applied. Currently available are
\code{"brier"} for the Brier score, \code{"log"} for the logarithmic score
and \code{"spherical"} for the spherical score.}
}
\value{
Vector of score values.
}
\description{
Computes consistent scores for forecast probabilities \code{p} and
observations \code{y}.
}
\details{
Smaller scores indicate better forecasts.
}
\references{
Preprint on e-values for forecast evaluation:

A. Henzi and J. F Ziegel. Valid sequential inference on probability forecast
performance. arXiv preprint arXiv:2103.08402, 2021.

Proper scoring rules for probabilities:

Gneiting, Tilmann, and Adrian E. Raftery. "Strictly proper scoring rules,
prediction, and estimation." Journal of the American statistical Association
102.477 (2007): 359-378.
}
