% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim1}
\alias{sim1}
\title{Simulation example 1}
\usage{
sim1(n, mu, type)
}
\arguments{
\item{n}{sample size (single integer).}

\item{mu}{parameter mu (single number in (0,1), see Details).}

\item{type}{scoring function. Currently available are \code{"brier"} for the
Brier score, \code{"log"} for the logarithmic score, \code{"spherical"}
for the spherical score.}
}
\value{
A list of forecasts \code{p}, \code{q} and observations \code{y}, with
attribute \code{type}.
}
\description{
Simulation example in Section 4.1 of the article.
}
\details{
Returns two simulated competing forecasts \code{p}, \code{q} and observations
\code{y}. The forecasts are random numbers in (0,1). If \code{type} is
\code{"brier"}, then the \code{y} take value 1 with probability
\code{mu*q+(1-mu)*p}. For \code{type} equal to \code{"log"} or
\code{"spherical"}, \code{\link{boundary}} is called to find the number
\code{r} such that \code{p} and \code{q} are equally good when the true
probability is \code{r}. The probability that \code{y} equals 1 is then
\code{r+(2mu-1)(q-r)} for \code{mu > 0.5} or \code{r+(1-2mu)(p-r)}
if \code{mu <= 0.5}. The forecast \code{p} is better than \code{q} if and
only if \code{mu <= 0.5}.
}
\references{
Preprint on e-values for forecast evaluation:

Proper scoring rules for probabilities:

Gneiting, Tilmann, and Adrian E. Raftery. "Strictly proper scoring rules,
prediction, and estimation." Journal of the American statistical Association
102.477 (2007): 359-378.
}
