% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim2}
\alias{sim2}
\title{Simulation example 2}
\usage{
sim2(n, theta, h)
}
\arguments{
\item{n}{sample size (single integer).}

\item{theta}{parameter theta (single number in (0,1)).}

\item{h}{maximal forecast lag (single integer).}
}
\value{
A list containing observations (\code{y}) and lagged forecasts (\code{pi_j}).
}
\description{
Simulation example in Section 4.2 of the article.
}
\details{
See Section 4.2 in "" for a description of the output.
}
\references{
Preprint on e-values for forecast evaluation:
}
